connected cloudinary to upload pictures as admin

+++ pages/api/products/[id].js
@@ -5,7 +5,10 @@ export default async function handler(req, res) {
     const { 
         method, 
         query: { id },
+        cookies,
      } = req;
+
+    const token = cookies.token
 
     await dbConnect()
 
@@ -19,6 +22,10 @@ export default async function handler(req, res) {
     }
 
     if (method === 'PUT') {
+        if(!token || token !== process.env.TOKEN) {
+            return res.status(401).json('You are not authenticated')
+        }
+
         try {
             const product = await Product.create(req.body);
             res.status(201).json(product);
@@ -28,6 +35,10 @@ export default async function handler(req, res) {
         }
     }
     if (method === 'DELETE') {
+        if(!token || token !== process.env.TOKEN) {
+            return res.status(401).json('You are not authenticated')
+        }
+        
         try {
             await Product.findByIdAndDelete(id);
             res.status(201).json("The product has been successfully deleted!");
